HOTFIX: refactored project to use NPM scopes and lerna default naming
