+++ jest.config.js
@@ -1,7 +1,7 @@
 // For details, check: https://basarat.gitbooks.io/typescript/docs/testing/jest.html
 module.exports = {
   "roots": [
-    "<rootDir>/modules"
+    "<rootDir>/packages"
   ],
   "transform": {
     "^.+\\.tsx?$": "ts-jest"
