+++ lerna.json
@@ -5,7 +5,7 @@
     }
   },
   "packages": [
-    "modules/*"
+    "packages/*"
   ],
   "version": "0.0.0"
 }
