+++ packages/cli/src/__tests__/6ar-cli.test.ts
